sketch2life/requirements.txt
Flask==2.1.1
gunicorn==20.1.0
Pillow==8.3.2
numpy==1.24.3
tensorflow==2.13.0
